requirements.txt
boto==2.49.0
boto3==1.9.191
botocore==1.12.191
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
django-heroku==0.3.1
django-pure-pagination==0.3.0
django-sass-processor==0.7.3
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
libsass==0.19.2
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
whitenoise==4.1.3